requirements.txt
python-telegram-bot[webhooks]==21.4
httpx==0.27.0
requests
python-dotenv
